+++ connection.go
@@ -18,6 +18,8 @@ import (
 	"sync"
 	"sync/atomic"
 	"time"
+
+	"github.com/vmihailenco/msgpack/v5"
 )
 
 const requestsMap = 128
@@ -154,7 +156,7 @@ type Connection struct {
 	rlimit  chan struct{}
 	opts    Opts
 	state   uint32
-	dec     *decoder
+	dec     *msgpack.Decoder
 	lenbuf  [PacketLengthBytes]byte
 
 	lastStreamId uint64
@@ -215,7 +217,7 @@ type connShard struct {
 	requestsWithCtx [requestsMap]futureList
 	bufmut          sync.Mutex
 	buf             smallWBuf
-	enc             *encoder
+	enc             *msgpack.Encoder
 }
 
 // Greeting is a message sent by Tarantool on connect.
@@ -350,7 +352,7 @@ func Connect(addr string, opts Opts) (conn *Connection, err error) {
 		Greeting:         &Greeting{},
 		control:          make(chan struct{}),
 		opts:             opts.Clone(),
-		dec:              newDecoder(&smallBuf{}),
+		dec:              msgpack.NewDecoder(&smallBuf{}),
 	}
 	maxprocs := uint32(runtime.GOMAXPROCS(-1))
 	if conn.opts.Concurrency == 0 || conn.opts.Concurrency > maxprocs*128 {
@@ -596,7 +598,7 @@ func (conn *Connection) dial() (err error) {
 	return
 }
 
-func pack(h *smallWBuf, enc *encoder, reqid uint32,
+func pack(h *smallWBuf, enc *msgpack.Encoder, reqid uint32,
 	req Request, streamId uint64, res SchemaResolver) (err error) {
 	const uint32Code = 0xce
 	const uint64Code = 0xcf
@@ -648,7 +650,7 @@ func pack(h *smallWBuf, enc *encoder, reqid uint32,
 
 func (conn *Connection) writeRequest(w *bufio.Writer, req Request) error {
 	var packet smallWBuf
-	err := pack(&packet, newEncoder(&packet), 0, req, ignoreStreamId, conn.Schema)
+	err := pack(&packet, msgpack.NewEncoder(&packet), 0, req, ignoreStreamId, conn.Schema)
 
 	if err != nil {
 		return fmt.Errorf("pack error: %w", err)
@@ -882,7 +884,7 @@ func (conn *Connection) writer(w *bufio.Writer, c net.Conn) {
 func readWatchEvent(reader io.Reader) (connWatchEvent, error) {
 	keyExist := false
 	event := connWatchEvent{}
-	d := newDecoder(reader)
+	d := msgpack.NewDecoder(reader)
 
 	l, err := d.DecodeMapLen()
 	if err != nil {
@@ -1114,7 +1116,7 @@ func (conn *Connection) putFuture(fut *Future, req Request, streamId uint64) {
 	firstWritten := shard.buf.Len() == 0
 	if shard.buf.Cap() == 0 {
 		shard.buf.b = make([]byte, 0, 128)
-		shard.enc = newEncoder(&shard.buf)
+		shard.enc = msgpack.NewEncoder(&shard.buf)
 	}
 	blen := shard.buf.Len()
 	reqid := fut.requestId