+++ request.go
@@ -6,91 +6,93 @@ import (
 	"reflect"
 	"strings"
 	"sync"
+
+	"github.com/vmihailenco/msgpack/v5"
 )
 
-func fillSearch(enc *encoder, spaceNo, indexNo uint32, key interface{}) error {
-	encodeUint(enc, KeySpaceNo)
-	encodeUint(enc, uint64(spaceNo))
-	encodeUint(enc, KeyIndexNo)
-	encodeUint(enc, uint64(indexNo))
-	encodeUint(enc, KeyKey)
+func fillSearch(enc *msgpack.Encoder, spaceNo, indexNo uint32, key interface{}) error {
+	enc.EncodeUint(KeySpaceNo)
+	enc.EncodeUint(uint64(spaceNo))
+	enc.EncodeUint(KeyIndexNo)
+	enc.EncodeUint(uint64(indexNo))
+	enc.EncodeUint(KeyKey)
 	return enc.Encode(key)
 }
 
-func fillIterator(enc *encoder, offset, limit, iterator uint32) {
-	encodeUint(enc, KeyIterator)
-	encodeUint(enc, uint64(iterator))
-	encodeUint(enc, KeyOffset)
-	encodeUint(enc, uint64(offset))
-	encodeUint(enc, KeyLimit)
-	encodeUint(enc, uint64(limit))
+func fillIterator(enc *msgpack.Encoder, offset, limit, iterator uint32) {
+	enc.EncodeUint(KeyIterator)
+	enc.EncodeUint(uint64(iterator))
+	enc.EncodeUint(KeyOffset)
+	enc.EncodeUint(uint64(offset))
+	enc.EncodeUint(KeyLimit)
+	enc.EncodeUint(uint64(limit))
 }
 
-func fillInsert(enc *encoder, spaceNo uint32, tuple interface{}) error {
+func fillInsert(enc *msgpack.Encoder, spaceNo uint32, tuple interface{}) error {
 	enc.EncodeMapLen(2)
-	encodeUint(enc, KeySpaceNo)
-	encodeUint(enc, uint64(spaceNo))
-	encodeUint(enc, KeyTuple)
+	enc.EncodeUint(KeySpaceNo)
+	enc.EncodeUint(uint64(spaceNo))
+	enc.EncodeUint(KeyTuple)
 	return enc.Encode(tuple)
 }
 
-func fillSelect(enc *encoder, spaceNo, indexNo, offset, limit, iterator uint32, key interface{}) error {
+func fillSelect(enc *msgpack.Encoder, spaceNo, indexNo, offset, limit, iterator uint32, key interface{}) error {
 	enc.EncodeMapLen(6)
 	fillIterator(enc, offset, limit, iterator)
 	return fillSearch(enc, spaceNo, indexNo, key)
 }
 
-func fillUpdate(enc *encoder, spaceNo, indexNo uint32, key, ops interface{}) error {
+func fillUpdate(enc *msgpack.Encoder, spaceNo, indexNo uint32, key, ops interface{}) error {
 	enc.EncodeMapLen(4)
 	if err := fillSearch(enc, spaceNo, indexNo, key); err != nil {
 		return err
 	}
-	encodeUint(enc, KeyTuple)
+	enc.EncodeUint(KeyTuple)
 	return enc.Encode(ops)
 }
 
-func fillUpsert(enc *encoder, spaceNo uint32, tuple, ops interface{}) error {
+func fillUpsert(enc *msgpack.Encoder, spaceNo uint32, tuple, ops interface{}) error {
 	enc.EncodeMapLen(3)
-	encodeUint(enc, KeySpaceNo)
-	encodeUint(enc, uint64(spaceNo))
-	encodeUint(enc, KeyTuple)
+	enc.EncodeUint(KeySpaceNo)
+	enc.EncodeUint(uint64(spaceNo))
+	enc.EncodeUint(KeyTuple)
 	if err := enc.Encode(tuple); err != nil {
 		return err
 	}
-	encodeUint(enc, KeyDefTuple)
+	enc.EncodeUint(KeyDefTuple)
 	return enc.Encode(ops)
 }
 
-func fillDelete(enc *encoder, spaceNo, indexNo uint32, key interface{}) error {
+func fillDelete(enc *msgpack.Encoder, spaceNo, indexNo uint32, key interface{}) error {
 	enc.EncodeMapLen(3)
 	return fillSearch(enc, spaceNo, indexNo, key)
 }
 
-func fillCall(enc *encoder, functionName string, args interface{}) error {
+func fillCall(enc *msgpack.Encoder, functionName string, args interface{}) error {
 	enc.EncodeMapLen(2)
-	encodeUint(enc, KeyFunctionName)
+	enc.EncodeUint(KeyFunctionName)
 	enc.EncodeString(functionName)
-	encodeUint(enc, KeyTuple)
+	enc.EncodeUint(KeyTuple)
 	return enc.Encode(args)
 }
 
-func fillEval(enc *encoder, expr string, args interface{}) error {
+func fillEval(enc *msgpack.Encoder, expr string, args interface{}) error {
 	enc.EncodeMapLen(2)
-	encodeUint(enc, KeyExpression)
+	enc.EncodeUint(KeyExpression)
 	enc.EncodeString(expr)
-	encodeUint(enc, KeyTuple)
+	enc.EncodeUint(KeyTuple)
 	return enc.Encode(args)
 }
 
-func fillExecute(enc *encoder, expr string, args interface{}) error {
+func fillExecute(enc *msgpack.Encoder, expr string, args interface{}) error {
 	enc.EncodeMapLen(2)
-	encodeUint(enc, KeySQLText)
+	enc.EncodeUint(KeySQLText)
 	enc.EncodeString(expr)
-	encodeUint(enc, KeySQLBind)
+	enc.EncodeUint(KeySQLBind)
 	return encodeSQLBind(enc, args)
 }
 
-func fillPing(enc *encoder) error {
+func fillPing(enc *msgpack.Encoder) error {
 	return enc.EncodeMapLen(0)
 }
 
@@ -195,7 +197,7 @@ type single struct {
 	found bool
 }
 
-func (s *single) DecodeMsgpack(d *decoder) error {
+func (s *single) DecodeMsgpack(d *msgpack.Decoder) error {
 	var err error
 	var len int
 	if len, err = d.DecodeArrayLen(); err != nil {
@@ -402,7 +404,7 @@ type KeyValueBind struct {
 // to avoid extra allocations in heap by calling strings.ToLower()
 var lowerCaseNames sync.Map
 
-func encodeSQLBind(enc *encoder, from interface{}) error {
+func encodeSQLBind(enc *msgpack.Encoder, from interface{}) error {
 	// internal function for encoding single map in msgpack
 	encodeKeyInterface := func(key string, val interface{}) error {
 		if err := enc.EncodeMapLen(1); err != nil {
@@ -534,8 +536,8 @@ func encodeSQLBind(enc *encoder, from interface{}) error {
 type Request interface {
 	// Code returns a IPROTO code for the request.
 	Code() int32
-	// Body fills an encoder with a request body.
-	Body(resolver SchemaResolver, enc *encoder) error
+	// Body fills an msgpack.Encoder with a request body.
+	Body(resolver SchemaResolver, enc *msgpack.Encoder) error
 	// Ctx returns a context of the request.
 	Ctx() context.Context
 	// Async returns true if the request does not expect response.
@@ -602,8 +604,8 @@ func newAuthRequest(user, scramble string) *authRequest {
 	return req
 }
 
-// Body fills an encoder with the auth request body.
-func (req *authRequest) Body(res SchemaResolver, enc *encoder) error {
+// Body fills an msgpack.Encoder with the auth request body.
+func (req *authRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 	return enc.Encode(map[uint32]interface{}{
 		KeyUserName: req.user,
 		KeyTuple:    []interface{}{string("chap-sha1"), string(req.scramble)},
@@ -623,8 +625,8 @@ func NewPingRequest() *PingRequest {
 	return req
 }
 
-// Body fills an encoder with the ping request body.
-func (req *PingRequest) Body(res SchemaResolver, enc *encoder) error {
+// Body fills an msgpack.Encoder with the ping request body.
+func (req *PingRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 	return fillPing(enc)
 }
 
@@ -699,8 +701,8 @@ func (req *SelectRequest) Key(key interface{}) *SelectRequest {
 	return req
 }
 
-// Body fills an encoder with the select request body.
-func (req *SelectRequest) Body(res SchemaResolver, enc *encoder) error {
+// Body fills an msgpack.Encoder with the select request body.
+func (req *SelectRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 	spaceNo, indexNo, err := res.ResolveSpaceIndex(req.space, req.index)
 	if err != nil {
 		return err
@@ -743,8 +745,8 @@ func (req *InsertRequest) Tuple(tuple interface{}) *InsertRequest {
 	return req
 }
 
-// Body fills an encoder with the insert request body.
-func (req *InsertRequest) Body(res SchemaResolver, enc *encoder) error {
+// Body fills an msgpack.Encoder with the insert request body.
+func (req *InsertRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 	spaceNo, _, err := res.ResolveSpaceIndex(req.space, nil)
 	if err != nil {
 		return err
@@ -787,8 +789,8 @@ func (req *ReplaceRequest) Tuple(tuple interface{}) *ReplaceRequest {
 	return req
 }
 
-// Body fills an encoder with the replace request body.
-func (req *ReplaceRequest) Body(res SchemaResolver, enc *encoder) error {
+// Body fills an msgpack.Encoder with the replace request body.
+func (req *ReplaceRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 	spaceNo, _, err := res.ResolveSpaceIndex(req.space, nil)
 	if err != nil {
 		return err
@@ -838,8 +840,8 @@ func (req *DeleteRequest) Key(key interface{}) *DeleteRequest {
 	return req
 }
 
-// Body fills an encoder with the delete request body.
-func (req *DeleteRequest) Body(res SchemaResolver, enc *encoder) error {
+// Body fills an msgpack.Encoder with the delete request body.
+func (req *DeleteRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 	spaceNo, indexNo, err := res.ResolveSpaceIndex(req.space, req.index)
 	if err != nil {
 		return err
@@ -900,8 +902,8 @@ func (req *UpdateRequest) Operations(ops *Operations) *UpdateRequest {
 	return req
 }
 
-// Body fills an encoder with the update request body.
-func (req *UpdateRequest) Body(res SchemaResolver, enc *encoder) error {
+// Body fills an msgpack.Encoder with the update request body.
+func (req *UpdateRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 	spaceNo, indexNo, err := res.ResolveSpaceIndex(req.space, req.index)
 	if err != nil {
 		return err
@@ -955,8 +957,8 @@ func (req *UpsertRequest) Operations(ops *Operations) *UpsertRequest {
 	return req
 }
 
-// Body fills an encoder with the upsert request body.
-func (req *UpsertRequest) Body(res SchemaResolver, enc *encoder) error {
+// Body fills an msgpack.Encoder with the upsert request body.
+func (req *UpsertRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 	spaceNo, _, err := res.ResolveSpaceIndex(req.space, nil)
 	if err != nil {
 		return err
@@ -1002,8 +1004,8 @@ func (req *CallRequest) Args(args interface{}) *CallRequest {
 	return req
 }
 
-// Body fills an encoder with the call request body.
-func (req *CallRequest) Body(res SchemaResolver, enc *encoder) error {
+// Body fills an msgpack.Encoder with the call request body.
+func (req *CallRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 	return fillCall(enc, req.function, req.args)
 }
 
@@ -1059,8 +1061,8 @@ func (req *EvalRequest) Args(args interface{}) *EvalRequest {
 	return req
 }
 
-// Body fills an encoder with the eval request body.
-func (req *EvalRequest) Body(res SchemaResolver, enc *encoder) error {
+// Body fills an msgpack.Encoder with the eval request body.
+func (req *EvalRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 	return fillEval(enc, req.expr, req.args)
 }
 
@@ -1099,8 +1101,8 @@ func (req *ExecuteRequest) Args(args interface{}) *ExecuteRequest {
 	return req
 }
 
-// Body fills an encoder with the execute request body.
-func (req *ExecuteRequest) Body(res SchemaResolver, enc *encoder) error {
+// Body fills an msgpack.Encoder with the execute request body.
+func (req *ExecuteRequest) Body(res SchemaResolver, enc *msgpack.Encoder) error {
 	return fillExecute(enc, req.expr, req.args)
 }