+++ request_test.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/vmihailenco/msgpack/v5"
 
 	. "github.com/tarantool/go-tarantool/v2"
 )
@@ -61,7 +62,7 @@ func assertBodyCall(t testing.TB, requests []Request, errorMsg string) {
 	const errBegin = "An unexpected Request.Body() "
 	for _, req := range requests {
 		var reqBuf bytes.Buffer
-		enc := NewEncoder(&reqBuf)
+		enc := msgpack.NewEncoder(&reqBuf)
 
 		err := req.Body(&resolver, enc)
 		if err != nil && errorMsg != "" && err.Error() != errorMsg {
@@ -80,7 +81,7 @@ func assertBodyEqual(t testing.TB, reference []byte, req Request) {
 	t.Helper()
 
 	var reqBuf bytes.Buffer
-	reqEnc := NewEncoder(&reqBuf)
+	reqEnc := msgpack.NewEncoder(&reqBuf)
 
 	err := req.Body(&resolver, reqEnc)
 	if err != nil {
@@ -229,7 +230,7 @@ func TestRequestsAsync(t *testing.T) {
 func TestPingRequestDefaultValues(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplPingBody(refEnc)
 	if err != nil {
 		t.Errorf("An unexpected RefImplPingBody() error: %q", err.Error())
@@ -243,7 +244,7 @@ func TestPingRequestDefaultValues(t *testing.T) {
 func TestSelectRequestDefaultValues(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplSelectBody(refEnc, validSpace, defaultIndex, 0, 0xFFFFFFFF, IterAll, []interface{}{})
 	if err != nil {
 		t.Errorf("An unexpected RefImplSelectBody() error %q", err.Error())
@@ -258,7 +259,7 @@ func TestSelectRequestDefaultIteratorEqIfKey(t *testing.T) {
 	var refBuf bytes.Buffer
 	key := []interface{}{uint(18)}
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplSelectBody(refEnc, validSpace, defaultIndex, 0, 0xFFFFFFFF, IterEq, key)
 	if err != nil {
 		t.Errorf("An unexpected RefImplSelectBody() error %q", err.Error())
@@ -275,7 +276,7 @@ func TestSelectRequestIteratorNotChangedIfKey(t *testing.T) {
 	key := []interface{}{uint(678)}
 	const iter = IterGe
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplSelectBody(refEnc, validSpace, defaultIndex, 0, 0xFFFFFFFF, iter, key)
 	if err != nil {
 		t.Errorf("An unexpected RefImplSelectBody() error %q", err.Error())
@@ -295,7 +296,7 @@ func TestSelectRequestSetters(t *testing.T) {
 	key := []interface{}{uint(36)}
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplSelectBody(refEnc, validSpace, validIndex, offset, limit, iter, key)
 	if err != nil {
 		t.Errorf("An unexpected RefImplSelectBody() error %q", err.Error())
@@ -314,7 +315,7 @@ func TestSelectRequestSetters(t *testing.T) {
 func TestInsertRequestDefaultValues(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplInsertBody(refEnc, validSpace, []interface{}{})
 	if err != nil {
 		t.Errorf("An unexpected RefImplInsertBody() error: %q", err.Error())
@@ -329,7 +330,7 @@ func TestInsertRequestSetters(t *testing.T) {
 	tuple := []interface{}{uint(24)}
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplInsertBody(refEnc, validSpace, tuple)
 	if err != nil {
 		t.Errorf("An unexpected RefImplInsertBody() error: %q", err.Error())
@@ -344,7 +345,7 @@ func TestInsertRequestSetters(t *testing.T) {
 func TestReplaceRequestDefaultValues(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplReplaceBody(refEnc, validSpace, []interface{}{})
 	if err != nil {
 		t.Errorf("An unexpected RefImplReplaceBody() error: %q", err.Error())
@@ -359,7 +360,7 @@ func TestReplaceRequestSetters(t *testing.T) {
 	tuple := []interface{}{uint(99)}
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplReplaceBody(refEnc, validSpace, tuple)
 	if err != nil {
 		t.Errorf("An unexpected RefImplReplaceBody() error: %q", err.Error())
@@ -374,7 +375,7 @@ func TestReplaceRequestSetters(t *testing.T) {
 func TestDeleteRequestDefaultValues(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplDeleteBody(refEnc, validSpace, defaultIndex, []interface{}{})
 	if err != nil {
 		t.Errorf("An unexpected RefImplDeleteBody() error: %q", err.Error())
@@ -389,7 +390,7 @@ func TestDeleteRequestSetters(t *testing.T) {
 	key := []interface{}{uint(923)}
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplDeleteBody(refEnc, validSpace, validIndex, key)
 	if err != nil {
 		t.Errorf("An unexpected RefImplDeleteBody() error: %q", err.Error())
@@ -405,7 +406,7 @@ func TestDeleteRequestSetters(t *testing.T) {
 func TestUpdateRequestDefaultValues(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplUpdateBody(refEnc, validSpace, defaultIndex, []interface{}{}, []Op{})
 	if err != nil {
 		t.Errorf("An unexpected RefImplUpdateBody() error: %q", err.Error())
@@ -421,7 +422,7 @@ func TestUpdateRequestSetters(t *testing.T) {
 	refOps, reqOps := getTestOps()
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplUpdateBody(refEnc, validSpace, validIndex, key, refOps)
 	if err != nil {
 		t.Errorf("An unexpected RefImplUpdateBody() error: %q", err.Error())
@@ -438,7 +439,7 @@ func TestUpdateRequestSetters(t *testing.T) {
 func TestUpsertRequestDefaultValues(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplUpsertBody(refEnc, validSpace, []interface{}{}, []Op{})
 	if err != nil {
 		t.Errorf("An unexpected RefImplUpsertBody() error: %q", err.Error())
@@ -454,7 +455,7 @@ func TestUpsertRequestSetters(t *testing.T) {
 	refOps, reqOps := getTestOps()
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplUpsertBody(refEnc, validSpace, tuple, refOps)
 	if err != nil {
 		t.Errorf("An unexpected RefImplUpsertBody() error: %q", err.Error())
@@ -470,7 +471,7 @@ func TestUpsertRequestSetters(t *testing.T) {
 func TestCallRequestsDefaultValues(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplCallBody(refEnc, validExpr, []interface{}{})
 	if err != nil {
 		t.Errorf("An unexpected RefImplCallBody() error: %q", err.Error())
@@ -489,7 +490,7 @@ func TestCallRequestsSetters(t *testing.T) {
 	args := []interface{}{uint(34)}
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplCallBody(refEnc, validExpr, args)
 	if err != nil {
 		t.Errorf("An unexpected RefImplCallBody() error: %q", err.Error())
@@ -510,7 +511,7 @@ func TestCallRequestsSetters(t *testing.T) {
 func TestEvalRequestDefaultValues(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplEvalBody(refEnc, validExpr, []interface{}{})
 	if err != nil {
 		t.Errorf("An unexpected RefImplEvalBody() error: %q", err.Error())
@@ -525,7 +526,7 @@ func TestEvalRequestSetters(t *testing.T) {
 	args := []interface{}{uint(34), int(12)}
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplEvalBody(refEnc, validExpr, args)
 	if err != nil {
 		t.Errorf("An unexpected RefImplEvalBody() error: %q", err.Error())
@@ -540,7 +541,7 @@ func TestEvalRequestSetters(t *testing.T) {
 func TestExecuteRequestDefaultValues(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplExecuteBody(refEnc, validExpr, []interface{}{})
 	if err != nil {
 		t.Errorf("An unexpected RefImplExecuteBody() error: %q", err.Error())
@@ -555,7 +556,7 @@ func TestExecuteRequestSetters(t *testing.T) {
 	args := []interface{}{uint(11)}
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplExecuteBody(refEnc, validExpr, args)
 	if err != nil {
 		t.Errorf("An unexpected RefImplExecuteBody() error: %q", err.Error())
@@ -570,7 +571,7 @@ func TestExecuteRequestSetters(t *testing.T) {
 func TestPrepareRequestDefaultValues(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplPrepareBody(refEnc, validExpr)
 	if err != nil {
 		t.Errorf("An unexpected RefImplPrepareBody() error: %q", err.Error())
@@ -584,7 +585,7 @@ func TestPrepareRequestDefaultValues(t *testing.T) {
 func TestUnprepareRequestDefaultValues(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplUnprepareBody(refEnc, *validStmt)
 	if err != nil {
 		t.Errorf("An unexpected RefImplUnprepareBody() error: %q", err.Error())
@@ -600,7 +601,7 @@ func TestExecutePreparedRequestSetters(t *testing.T) {
 	args := []interface{}{uint(11)}
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplExecutePreparedBody(refEnc, *validStmt, args)
 	if err != nil {
 		t.Errorf("An unexpected RefImplExecutePreparedBody() error: %q", err.Error())
@@ -616,7 +617,7 @@ func TestExecutePreparedRequestSetters(t *testing.T) {
 func TestExecutePreparedRequestDefaultValues(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplExecutePreparedBody(refEnc, *validStmt, []interface{}{})
 	if err != nil {
 		t.Errorf("An unexpected RefImplExecutePreparedBody() error: %q", err.Error())
@@ -631,7 +632,7 @@ func TestExecutePreparedRequestDefaultValues(t *testing.T) {
 func TestBeginRequestDefaultValues(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplBeginBody(refEnc, defaultIsolationLevel, defaultTimeout)
 	if err != nil {
 		t.Errorf("An unexpected RefImplBeginBody() error: %q", err.Error())
@@ -645,7 +646,7 @@ func TestBeginRequestDefaultValues(t *testing.T) {
 func TestBeginRequestSetters(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplBeginBody(refEnc, ReadConfirmedLevel, validTimeout)
 	if err != nil {
 		t.Errorf("An unexpected RefImplBeginBody() error: %q", err.Error())
@@ -659,7 +660,7 @@ func TestBeginRequestSetters(t *testing.T) {
 func TestCommitRequestDefaultValues(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplCommitBody(refEnc)
 	if err != nil {
 		t.Errorf("An unexpected RefImplCommitBody() error: %q", err.Error())
@@ -673,7 +674,7 @@ func TestCommitRequestDefaultValues(t *testing.T) {
 func TestRollbackRequestDefaultValues(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	err := RefImplRollbackBody(refEnc)
 	if err != nil {
 		t.Errorf("An unexpected RefImplRollbackBody() error: %q", err.Error())
@@ -687,7 +688,7 @@ func TestRollbackRequestDefaultValues(t *testing.T) {
 func TestBroadcastRequestDefaultValues(t *testing.T) {
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	expectedArgs := []interface{}{validKey}
 	err := RefImplCallBody(refEnc, "box.broadcast", expectedArgs)
 	if err != nil {
@@ -703,7 +704,7 @@ func TestBroadcastRequestSetters(t *testing.T) {
 	value := []interface{}{uint(34), int(12)}
 	var refBuf bytes.Buffer
 
-	refEnc := NewEncoder(&refBuf)
+	refEnc := msgpack.NewEncoder(&refBuf)
 	expectedArgs := []interface{}{validKey, value}
 	err := RefImplCallBody(refEnc, "box.broadcast", expectedArgs)
 	if err != nil {