+++ tarantool_test.go
@@ -13,9 +13,10 @@ import (
 	"testing"
 	"time"
 
-	"github.com/stretchr/testify/require"
-
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"github.com/vmihailenco/msgpack/v5"
+
 	. "github.com/tarantool/go-tarantool/v2"
 	"github.com/tarantool/go-tarantool/v2/test_helpers"
 )
@@ -36,20 +37,20 @@ type Member struct {
 	Val   uint
 }
 
-func (m *Member) EncodeMsgpack(e *encoder) error {
+func (m *Member) EncodeMsgpack(e *msgpack.Encoder) error {
 	if err := e.EncodeArrayLen(2); err != nil {
 		return err
 	}
 	if err := e.EncodeString(m.Name); err != nil {
 		return err
 	}
-	if err := encodeUint(e, uint64(m.Val)); err != nil {
+	if err := e.EncodeUint(uint64(m.Val)); err != nil {
 		return err
 	}
 	return nil
 }
 
-func (m *Member) DecodeMsgpack(d *decoder) error {
+func (m *Member) DecodeMsgpack(d *msgpack.Decoder) error {
 	var err error
 	var l int
 	if l, err = d.DecodeArrayLen(); err != nil {